+++ packages/delir-core/src/Values/AssetPointer.ts
