+++ packages/delir-core/src/engine/Engine.ts
@@ -16,7 +16,7 @@ import defaults from '../helper/defaults'
 import FPSCounter from '../helper/FPSCounter'
 import ProgressPromise from '../helper/progress-promise'
 import * as ProjectHelper from '../helper/project-helper'
-import { ColorRGB, ColorRGBA } from '../values'
+import { ColorRGB, ColorRGBA } from '../Values'
 import AssetProxy from './AssetProxy'
 import DependencyResolver from './DependencyResolver'
 import * as ExpressionContext from './ExpressionSupport/ExpressionContext'
