+++ packages/delir-core/src/engine/Task/ClipRenderTask.ts
@@ -4,7 +4,7 @@ import * as KeyframeHelper from '../../helper/keyframe-helper'
 import { TypeDescriptor } from '../../plugin-support/type-descriptor'
 import { Clip } from '../../project'
 import { AssetPointerScheme } from '../../project/scheme/keyframe'
-import { Expression } from '../../values'
+import { Expression } from '../../Values'
 import { RealParameterValues, RealParameterValueTypes } from '../Engine'
 import { compileTypeScript } from '../ExpressionSupport/ExpressionCompiler'
 import * as ExpressionContext from '../ExpressionSupport/ExpressionContext'
