+++ packages/delir-core/src/engine/Task/EffectRenderTask.ts
@@ -5,7 +5,7 @@ import EffectPluginBase from '../../plugin-support/PostEffectBase'
 import { TypeDescriptor } from '../../plugin-support/type-descriptor'
 import { Clip, Effect } from '../../project'
 import { AssetPointerScheme } from '../../project/scheme/keyframe'
-import { Expression } from '../../values'
+import { Expression } from '../../Values'
 import DependencyResolver from '../DependencyResolver'
 import {  RealParameterValues, RealParameterValueTypes } from '../Engine'
 import { compileTypeScript } from '../ExpressionSupport/ExpressionCompiler'
