+++ packages/delir-core/src/helper/__spec__/project-helper-spec.ts
@@ -1,14 +1,14 @@
 // @flow
 import * as ProjectHelper from '../project-helper'
 
-import Project from '../../project/project'
 import Asset from '../../project/asset'
-import Composition from '../../project/composition'
-import Layer from '../../project/layer'
 import Clip from '../../project/clip'
+import Composition from '../../project/composition'
 import Keyframe from '../../project/keyframe'
+import Layer from '../../project/layer'
+import Project from '../../project/project'
 
-import ColorRGB from '../../values/color-rgb'
+import ColorRGB from '../../Values/color-rgb'
 
 const propNotWritable = (obj, prop) => {
     return Object.getOwnPropertyDescriptor(obj, prop).writable === false
