+++ packages/delir-core/src/project/keyframe.ts
@@ -2,13 +2,13 @@ import * as _ from 'lodash'
 import * as uuid from 'uuid'
 import toJSON from '../helper/toJSON'
 
-// import Point2D from '../values/point-2d'
-// import Point3D from '../values/point-3d'
-// import Size2D from '../values/size-2d'
-// import Size3D from '../values/size-3d'
-import ColorRGB from '../values/color-rgb'
-import ColorRGBA from '../values/color-rgba'
-import Expression from '../values/expression'
+// import Point2D from '../Values/point-2d'
+// import Point3D from '../Values/point-3d'
+// import Size2D from '../Values/size-2d'
+// import Size3D from '../Values/size-3d'
+import ColorRGB from '../Values/color-rgb'
+import ColorRGBA from '../Values/color-rgba'
+import Expression from '../Values/expression'
 
 import { AssetPointerScheme, JSONKeyframeValueTypes, KeyframeConfigScheme, KeyframeScheme } from './scheme/keyframe'
